Alterado app principal (pet_happy) para dois apps separados: pessoas e pets.
