+++ pessoas/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class PetHappyConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'pet_happy'
+    name = 'pessoas'
