+++ setup/urls.py
@@ -2,12 +2,12 @@ from django.contrib import admin
 from django.urls import path, include
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 
-from pet_happy.url import pet_happy_router
+from pessoas.urls import pessoas_router
 
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('token/', TokenObtainPairView.as_view()),
     path('token/refresh/', TokenRefreshView.as_view()),
-    path('', include(pet_happy_router.urls)),
+    path('', include(pessoas_router.urls)),
 ]
